added read books page and to-read option, updated previous models and controllers

+++ src/main/java/com/bjournal/bookjournal/controller/BookController.java
@@ -3,8 +3,11 @@ package com.bjournal.bookjournal.controller;
 import com.bjournal.bookjournal.model.Book;
 import com.bjournal.bookjournal.model.Review;
 import com.bjournal.bookjournal.model.UserReadBook;
+import com.bjournal.bookjournal.model.exceptions.BookNotFoundException;
+import com.bjournal.bookjournal.model.exceptions.UserNotAuthenticatedException;
 import com.bjournal.bookjournal.service.BookService;
 import com.bjournal.bookjournal.service.ReviewService;
+import com.bjournal.bookjournal.service.ToReadBookService;
 import com.bjournal.bookjournal.service.UserReadBookService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.stereotype.Controller;
@@ -12,7 +15,9 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -21,11 +26,13 @@ public class BookController {
     private final BookService bookService;
     private final ReviewService reviewService;
     private final UserReadBookService userReadBookService;
+    private final ToReadBookService toReadBookService;
 
-    public BookController(BookService bookService, ReviewService reviewService, UserReadBookService userReadBookService) {
+    public BookController(BookService bookService, ReviewService reviewService, UserReadBookService userReadBookService, ToReadBookService toReadBookService) {
         this.bookService = bookService;
         this.reviewService = reviewService;
         this.userReadBookService = userReadBookService;
+        this.toReadBookService = toReadBookService;
     }
 
     @GetMapping("/add")
@@ -36,18 +43,30 @@ public class BookController {
     @PostMapping("/add")
     public String addBook(@RequestParam(required = true) String title, @RequestParam(required = true) String author,
                           @RequestParam(required = true) String description, @RequestParam(required = true) MultipartFile file,
-                          @RequestParam(required = true) Integer pages) {
+                          @RequestParam(required = true) Integer pages, Model model) {
         String contentType = file.getContentType();
         if (!contentType.equals("image/jpeg") && !contentType.equals("image/png")) {
-            throw new IllegalArgumentException("Only JPEG or PNG images are allowed");
+            model.addAttribute("error", "Only JPEG or PNG images are allowed");
+            model.addAttribute("hasError", true);
+            return "book-form";
         }
-        this.bookService.add(title, author, description, file, pages);
+        try {
+            this.bookService.add(title, author, description, file, pages);
+        } catch (IllegalArgumentException e) {
+            model.addAttribute("error", e.getMessage());
+            model.addAttribute("hasError", true);
+            return "book-form";
+        } catch (IOException e) {
+            model.addAttribute("error", "Error while saving the image");
+            model.addAttribute("hasError", true);
+            return "book-form";
+        }
         return "redirect:/";
     }
 
     @GetMapping("/edit/{id}")
     public String editBook(@PathVariable Long id, Model model) {
-        if(this.bookService.findById(id).isPresent()){
+        if (this.bookService.findById(id).isPresent()) {
             model.addAttribute("book", bookService.findById(id).get());
             return "book-form";
         }
@@ -57,8 +76,26 @@ public class BookController {
     @PostMapping("/edit/{id}")
     public String editBook(@PathVariable Long id, @RequestParam(required = true) String title,
                            @RequestParam(required = true) String author, @RequestParam(required = true) String description,
-                           @RequestParam(required = true) MultipartFile file, @RequestParam(required = true) Integer pages) {
-        this.bookService.update(id, title, author, description, file, pages);
+                           @RequestParam(required = true) MultipartFile file, @RequestParam(required = true) Integer pages,
+                           Model model) {
+        try {
+            this.bookService.update(id, title, author, description, file, pages);
+        } catch (BookNotFoundException e) {
+            model.addAttribute("error", e.getMessage());
+            return "redirect:/?error=BookNotFound";
+        } catch (IllegalArgumentException e) {
+            Book book = bookService.findById(id).get();
+            model.addAttribute("book", book);
+            model.addAttribute("error", e.getMessage());
+            model.addAttribute("hasError", true);
+            return "book-form";
+        } catch (IOException e) {
+            Book book = bookService.findById(id).get();
+            model.addAttribute("book", book);
+            model.addAttribute("error", "Error while saving the image");
+            model.addAttribute("hasError", true);
+            return "book-form";
+        }
         return "redirect:/";
     }
 
@@ -70,13 +107,20 @@ public class BookController {
 
     @GetMapping("/details/{id}")
     public String detailsBook(@PathVariable Long id, Model model, HttpServletRequest req) {
-        if(this.bookService.findById(id).isPresent()) {
+        String username = req.getRemoteUser();
+        if (username == null) {
+            throw new UserNotAuthenticatedException();
+        }
+        if (this.bookService.findById(id).isPresent()) {
             Book book = this.bookService.findById(id).get();
             List<Review> reviews = this.reviewService.findAllByBook(book);
+            if (toReadBookService.findByUserAndBook(username, book).isPresent()) {
+                model.addAttribute("toReadBook", true);
+            } else {
+                model.addAttribute("toReadBook", false);
+            }
             model.addAttribute("book", book);
             model.addAttribute("reviews", reviews);
-            String username = req.getRemoteUser();
-
             if (userReadBookService.findByUserAndBook(username, book).isPresent()) {
                 UserReadBook hasReadBook = userReadBookService.findByUserAndBook(username, book).get();
                 model.addAttribute("hasRead", true);
@@ -92,9 +136,13 @@ public class BookController {
 
     @GetMapping("/read/{id}")
     public String readBook(@PathVariable Long id, HttpServletRequest req) {
-        if(this.bookService.findById(id).isPresent()) {
+        if (this.bookService.findById(id).isPresent()) {
             Book book = this.bookService.findById(id).get();
             String username = req.getRemoteUser();
+            if (username == null) {
+                throw new UserNotAuthenticatedException();
+            }
+
             this.userReadBookService.add(LocalDate.now(), username, book);
             return "redirect:/books/details/" + id;
         }
@@ -104,12 +152,16 @@ public class BookController {
     @PostMapping("/read/{id}")
     public String readBook(@PathVariable Long id, @RequestParam(required = false) LocalDate startedDate,
                            @RequestParam(required = false) LocalDate finishedDate, HttpServletRequest req) {
-        if(this.bookService.findById(id).isPresent()) {
+        if (this.bookService.findById(id).isPresent()) {
             Book book = this.bookService.findById(id).get();
             String username = req.getRemoteUser();
+            if (username == null) {
+                throw new UserNotAuthenticatedException();
+            }
+
             if (userReadBookService.findByUserAndBook(username, book).isPresent()) {
                 UserReadBook hasReadBook = userReadBookService.findByUserAndBook(username, book).get();
-                this.userReadBookService.update(hasReadBook, startedDate,finishedDate);
+                this.userReadBookService.update(hasReadBook, startedDate, finishedDate);
                 return "redirect:/books/details/" + id;
             } else {
                 return "redirect:/?error=BookNotFound";
@@ -120,10 +172,14 @@ public class BookController {
 
     @GetMapping("/remove-read/{id}")
     public String removeReadBook(@PathVariable Long id, HttpServletRequest req) {
-        if(this.bookService.findById(id).isPresent()) {
+        if (this.bookService.findById(id).isPresent()) {
             Book book = this.bookService.findById(id).get();
             String username = req.getRemoteUser();
-            this.userReadBookService.deleteByUserAndBook(username,book);
+            if (username == null) {
+                throw new UserNotAuthenticatedException();
+            }
+
+            this.userReadBookService.deleteByUserAndBook(username, book);
             return "redirect:/books/details/" + id;
         }
         return "redirect:/?error=BookNotFound";
@@ -131,10 +187,46 @@ public class BookController {
 
     @PostMapping("/review/{id}")
     public String reviewBook(@PathVariable Long id, @RequestParam(required = true) String review, HttpServletRequest req) {
-        if (this.bookService.findById(id).isPresent()){
+        if (this.bookService.findById(id).isPresent()) {
             String username = req.getRemoteUser();
+            if (username == null) {
+                throw new UserNotAuthenticatedException();
+            }
+
             this.reviewService.addReview(review, username, id);
             return "redirect:/books/details/" + id;
+        }
+        return "redirect:/?error=BookNotFound";
+    }
+
+    @GetMapping("/read")
+    public String listReadBooks(HttpServletRequest req, Model model, @RequestParam(required = false) String search) {
+        String username = req.getRemoteUser();
+        if (username == null) {
+            throw new UserNotAuthenticatedException();
+        }
+        List<UserReadBook> readBooks = new ArrayList<>();
+        if (search == null || search.isBlank()) {
+            readBooks = this.userReadBookService.findAllByUser(username);
+        } else {
+            readBooks = this.userReadBookService.findAllByUserAndBookTitleContainingIgnoreCase(username, search);
+        }
+        model.addAttribute("readBooks", readBooks);
+        model.addAttribute("search", search);
+        return "read-books";
+    }
+
+    @GetMapping("/to-read/{id}")
+    public String toReadBook(@PathVariable Long id, @RequestParam(required = true) String redirect, HttpServletRequest req) {
+        if (this.bookService.findById(id).isPresent()) {
+            Book book = this.bookService.findById(id).get();
+            String username = req.getRemoteUser();
+            if (username == null) {
+                throw new UserNotAuthenticatedException();
+            }
+
+            this.toReadBookService.toggleToRead(username,book);
+            return "redirect:"+redirect;
         }
         return "redirect:/?error=BookNotFound";
     }
